+++ android/src/main/kotlin/com/kishormainali/native_crash/NativeCrashException.kt
@@ -3,8 +3,8 @@ package com.kishormainali.native_crash
 import androidx.annotation.Keep
 
 @Keep
-class NativeCrashException internal constructor() :
-    RuntimeException("This is a crash caused by calling .crash() in Dart.") {
+class NativeCrashException internal constructor(override val message: String) :
+    RuntimeException(message) {
     init {
         fillInStackTrace()
     }