+++ lib/native_crash_method_channel.dart
@@ -10,7 +10,27 @@ class MethodChannelNativeCrash extends NativeCrashPlatform {
   final methodChannel = const MethodChannel('com.kishormainali.native_crash');
 
   @override
-  Future<void> crash() async {
-    return await methodChannel.invokeMethod('crash');
+  Future<void> crash([String? message]) async {
+    return await methodChannel.invokeMethod('crash', {
+      'message': message ??
+          'This message is thrown when crash() is called from dart side.',
+    });
+  }
+
+  @override
+  Future<bool> checkJailBreak({bool enableLogging = false}) async {
+    return await methodChannel.invokeMethod('checkJailBreak', {
+      'enableLogging': enableLogging,
+    });
+  }
+
+  @override
+  Future<bool> checkDevMode() async {
+    return await methodChannel.invokeMethod('isDevMode');
+  }
+
+  @override
+  Future<bool> checkEmulator() async {
+    return await methodChannel.invokeMethod('isEmulator');
   }
 }