+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+- added `jailbreak` or `root` detection.
+- added optional `message` parameter to crash() method.
+
 ## 1.0.0
 
 - initial release.